fix testcases

+++ UnitTest_Nunit_Triangle/Diem.cs
@@ -8,12 +8,12 @@ namespace UnitTest_Nunit_Triangle
 {
     public class Diem
     {
-        public double _x;
-        public double _y;
+        public decimal _x;
+        public decimal _y;
 
-        public void SetX(double value) => this._x = value;
-        public double GetX() => this._x;
-        public void SetY(double value) => this._y = value;
-        public double GetY() => this._y;
+        public void SetX(decimal value) => this._x = value;
+        public decimal GetX() => this._x;
+        public void SetY(decimal value) => this._y = value;
+        public decimal GetY() => this._y;
     }
 }
